refactor: consumer-repo-next - lib createReactContextStore
