+++ apps/consumer-repo-next/lib/state/react-context/create-react-context-store.demo.tsx
@@ -1,4 +1,4 @@
-import { getRandomIdWithSeed } from "../utils/random";
+import { getRandomIdWithSeed } from "../../utils/random";
 import { createReactContextStore } from "./create-react-context-store";
 
 // domain types
